requirements.txt
asgiref==3.4.1
astroid==2.9.1
autopep8==1.6.0
bandit==1.7.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django>=3.2.12,<3.3.0
django-cleanup==5.2.0
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-jalali==5.0.0
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
djangorestframework>=3.13.1,<3.14.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
jdatetime==3.7.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
model-bakery==1.3.3
oauthlib==3.1.1
packaging==21.3
pbr==5.8.0
phonenumbers==8.12.40
Pillow>=9.0.1,<9.1.0
platformdirs==2.4.1
psycopg2==2.9.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pygraphviz==1.7
PyJWT>=2.4.0,<2.5.0
pylint==2.12.2
pyparsing==3.0.6
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
smmap==5.0.0
sqlparse==0.4.2
stevedore==3.5.0
toml==0.10.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
